feat : 리뷰 생성 기능 구현

+++ src/main/java/sparta/ifour/movietalk/domain/reviews/dto/ReviewResponseDto.java
@@ -17,7 +17,7 @@ public class ReviewResponseDto {
 
     private String content;
 
-    private Double ratingScore;
+    private Float ratingScore;
 
     private String movieName;
 
@@ -32,7 +32,6 @@ public class ReviewResponseDto {
         this.movieName = review.getMovieName();
         this.createdAt = review.getCreatedAt();
     }
-
